pyproject.toml
[tool.poetry]
name = "audio_extractor"
version = "0.1.0"
description = ""
authors = ["Anatoliy Lazovyj <lazoviy3000@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
vkaudiotoken = "^0.5.1"
vk-api = "^11.9.9"
vkwave = {git = "https://github.com/Tolyan3k/vkwave"}
pyotp = "^2.9.0"
loguru = "^0.7.2"
ffmpeg-python = "^0.2.0"
yt-dlp = "^2024.3.10"
pathvalidate = "^3.2.0"
sqlmodel = "^0.0.16"
pydantic-settings = "^2.2.1"
tenacity = "^8.2.3"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
moviepy = "^1.0.3"

[tool.poetry.group.testing.dependencies]
pytest = "8.0.0"
pytest-cov = "^5.0.0"
pytest-xdist = {extras = ["psutil"], version = "^3.6.1"}

[tool.poetry.group.linting.dependencies]
pylint = "3.0.3"
ruff = "0.3.0"

[tool.poetry.group.formatting.dependencies]
black = "^24.2.0"
yapf = "^0.40.2"

[tool.poetry.group.profiling.dependencies]
scalene = {version = "^1.5.41", platform = "linux"}
memray = {version = "^1.12.0", platform = "linux"}


[tool.poetry.group.git.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
extend-exclude = "settings.py"  # Cannot parse line:
                                # default=f"sqlite:///{_BotWorkDirs().DIRS["data"]}/database.db",
